add subscription convenience helper and general cleanup
